B1-B4/B6 del playtest profondo: passaggio a proprietario unico, la fuga arretra, status a tabella
- B1: `passaggio_concesso()` in mappa = UNICO proprietario di «puoi lasciare
  la stanza» — compositore (componi_opzioni_scena), muovi() e la discesa
  leggono lo stesso predicato; il «Vai» in tregua muove davvero e la scala
  non è mai una fuga gratis. Lucchetto di COPPIA: ogni MUOVI composto muta
  stanza_corrente.
- B2: la fuga in combattimento riuscita È il disimpegno — `arretra()` (stessa
  min(adiacenze) dello Scappi di scena, ora condivisa) come parte della
  risoluzione + DisimpegnoScena in cronaca; lo Scappi fallito pubblica
  DisimpegnoFallito PRIMA dell'EncounterStarted (eventi.py:132 rivista:
  niente click muto). Chiuso il tritacarne di Evil Ash.
- B3: colonna `descrittore` in SPEC_STATUS; proietta_scheda itera la tabella
  (il Brucia si vede: «in fiamme») e main assorbe la terza copia della mappa
  slug→participio (DESCRITTORI_STATUS).
- B4: la difesa si proietta in unità di mitigazione (def_eff/100, 1 decimale;
  derivate: Mapping[str, float]) — DEF 238 era presentazione, non bilanciamento.
- B6: etichetta box e apertura leggono la STESSA `grado_apertura` (scalata
  mostrata: «Armi (Bronzo → Argento qui)»).
- B7.2 (motore): slot fissi del menu di scena — tempo al centro, riposo dopo
  l'attesa, movimento sempre in fondo.
- B7.4: ParlamentoRisolto sul bus (classe + esito) alla prova fresca del gate.
- B7.5: epitaffio coi round cumulati dell'avversario fatale su più ingaggi.

+++ src/contracts/proiezione.py
@@ -215,7 +215,12 @@ class SchedaVista(BaseModel):
     descrittori: tuple[str, ...] = ()
     primarie: Mapping[str, int] = {}
     primarie_occulte: tuple[str, ...] = ()
-    derivate: Mapping[str, int] = {}
+    # `float`, non `int` (B4, playtest profondo 2026-08-28): la difesa si
+    # proietta in UNITÀ DI MITIGAZIONE (`def_eff/100`, una cifra decimale) —
+    # esposta grezza in centesimi accanto a derivate in scala intera sembrava
+    # un bug di bilanciamento (DEF 238 con danni invariati). Le altre derivate
+    # restano interi (un `int` è un `float` valido per Pydantic e per JSON).
+    derivate: Mapping[str, float] = {}
     livello: int = 1
     tick_piano: int = 0
